Add support for importing .blend files
Lets you drag or place .blend files in the project folder and it will import the files.

An editor setting sets the location of the blender binary. 

Checks for Blender 3.0's gltf2 `export_keep_originals` option.
Add support for importing and exporting .blend files

Lets you drag or place .blend files in the project folder and it will import the files.

An editor setting sets the location of the blender binary. A empty setting means the feature is disabled.

Checks for Blender 3.0's gltf2 keep texture option.