Add AnimationTree Advance Expressions
This gives much greater flexibility for creating complex state machines. By directly interfacing with the script code, it is possible to create complex animation advance condition for switching between states.

Ensure assigning AnimationTreeStateMachineTransition base expression node in editor is relative to current AnimationTree node.

Allow setting an expression base node on the AnimationTree itself

Allows specifying an expression as a condition for state machine transitions.

For this to work safely (user not call queue_free or something in the expression), a const call mode was added to Object and Variant (and optionally Script).

Fixed an animation import bug in Collada.

Co-Authored-By: reduz <reduzio@gmail.com>