Remove setup_humanoid_bones method and related code
The setup_humanoid_bones method has been removed as it is no longer needed. This method was responsible for setting up the bones for a humanoid structure, but it has been decided to remove this functionality. The code related to this method has also been removed from the ManyBoneIK3D class.